requirements.txt
# ── Core ──────────────────────────────────────────────────────
boto3>=1.34.0
botocore>=1.34.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.1.0

# ── RL & ML ───────────────────────────────────────────────────
torch>=2.2.0
stable-baselines3>=2.2.0
gymnasium>=0.29.0
mlflow>=2.10.0
h5py>=3.10.0
scikit-learn>=1.4.0

# ── Async / Orchestration ─────────────────────────────────────
asyncio-throttle>=1.0.2
aiohttp>=3.9.0
tenacity>=8.2.3         # retry logic for AWS API calls

# ── API (Dashboard Backend) ───────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic>=2.6.0

# ── Storage / DB ──────────────────────────────────────────────
aioboto3>=12.3.0        # async boto3 for DynamoDB writes

# ── Monitoring / Observability ────────────────────────────────
prometheus-client>=0.20.0
rich>=13.7.0            # pretty CLI output

# ── Testing ───────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
moto[ec2,s3,dynamodb]>=5.0.0   # mock AWS for tests
